Fix for nasty Heisenbug(tm): C++ enums can only be used as flag values when the variable is of type int.
With short etc. any concatenation ("|" operator) of flags will generate
an int, and in this case was overwriting neighboring fields in
OperationDepsNode ("flag" and "done" fields), which was causing havoc
in the scheduler.

Alternatively a specialized flag type can be defined. See here for
further info on the issue:
http://stackoverflow.com/questions/1448396/how-to-use-enums-as-flags-in-c

+++ source/blender/depsgraph/intern/depsgraph_eval.h
@@ -56,7 +56,7 @@ typedef struct DEG_OperationsContext {
 	
 	int type;             /* (eDepsNode_Type.OuterNodes) component type <-> context type (for debug purposes) */
 	short utype;          /* (eDEG_OperationContext_UserType) evaluation user type */
-	short flag;           /* (eDEG_OperationContext_Flag) extra settings */
+	int flag;             /* (eDEG_OperationContext_Flag) extra settings */
 } DEG_OperationsContext;
 
 /* Settings */
